baitap-submit/dan-dinh/05-diffusion-gradio/requirements.txt
groq==0.15.0
python-dotenv==0.21.0
gradio==5.12.0
httpx==0.27.2
diffusers==0.32.2
torch==2.5.1+cu124